search_algos.h

+++ 0x1E-search_algorithms/search_algos.h
@@ -10,6 +10,6 @@ int binary_search(int *array, size_t size, int value);
 
 /* helper functions*/
 int binary_search_recurs(int *array, size_t lb, size_t ub, int value);
-void print_array(int *array, size_t lb, size_t ub);
+void print_array(int *array, int lb, int ub);
 
 #endif /*SEARCH_ALGOS_HEADER*/
